Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_allow_script-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://localhost:8800/html/semantics/embedded-content/the-iframe-element/support/sandbox_allow_script.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.

PASS iframe_sandbox_allow_scripts
